requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
SQLAlchemy==2.0.34
pydantic-settings==2.8.1
python-dotenv==0.21.0

# --- LangChain stack (2025-Q2) ---
langchain==0.2.*
langchain-core==0.2.*
langchain-community==0.2.*
langchain-google-genai==0.1.*
langchain-huggingface==0.0.11

numpy==1.26.4
python-multipart==0.0.9
aiofiles==23.1.0 